+++ R/utils-git.R
@@ -1,5 +1,10 @@
+has_git <- function() {
+  Sys.which("git") != ""
+}
+
 # Shamelessly stolen from {pkgdown}, originally authored by Hadley Wickam
 git <- function (..., echo_cmd = TRUE, echo = TRUE, error_on_status = TRUE) {
+  if (!has_git()) stop(cli::format_error("{.pkg git} is not installed"), call. = FALSE)
   callr::run("git", c(...), echo_cmd = echo_cmd, echo = echo, 
     error_on_status = error_on_status)
 }
@@ -13,53 +18,144 @@ git_has_remote_branch <- function (remote, branch) {
     )$status == 0
 }
 
-# worktree setup
-# [IF BRANCH DOES NOT EXIST]
-#   git checkout --orphan <branch>
-#   git rm -rf --quiet .
-#   git commit --allow-empty -m
-#   git push remote HEAD:<branch>
-#   git checkout -
-# git remote set-branches <remote> <branch>
-# git fetch <remote> <branch>
-# git worktree add --track -B <branch> /path/to/dir <remote>/<branch>
+make_refspec <- function(remote, branch) {
+  glue::glue("+refs/heads/{branch}:refs/remotes/{remote}/{branch}")
+}
+
 #
 # Modified from pkgdown::deploy_to_branch() by Hadley Wickham
-git_worktree_setup <- function (path = ".", dest_dir, branch = "gh-pages", remote = "origin") {
 
-  no_branch <- !git_has_remote_branch(remote, branch)
+#' Setup a git worktree for concurrent manipulation of a separate branch
+#'
+#' @param path path to the repository
+#' @param dest_dir path to the destination directory to contain the work tree
+#' @param branch the branch associated with the work tree (default: gh-pages)
+#' @param remote the remote name (default: origin)
+#' @param throwaway if `TRUE`, the worktree created is in a detached HEAD state
+#'   from from the remote branch and will not create a new branch in your 
+#'   repository. Defaults to `FALSE`, which will create the branch from upstream.
+#' @return an [expression()] that calls `git worktree remove` on the worktree
+#'   when evaluated. 
+#' @details
+#'
+#' This function is used in continuous integration settings where we want to
+#' push derived outputs to non-main branches in our repository. We use this to
+#' populate the markdown and HTML outputs from the lesson so that we don't have
+#' to rebuild the lesson from scratch every time. 
+#'
+#' The logic behind this looks like
+#'
+#' ```
+#' worktree setup
+#' [IF BRANCH DOES NOT EXIST]
+#'   git checkout --orphan <branch>
+#'   git rm -rf --quiet .
+#'   git commit --allow-empty -m
+#'   git push remote HEAD:<branch>
+#'   git checkout -
+#' git fetch <remote> +refs/heads/<branch>:refs/remotes/<remote>/<branch>
+#' git worktree add --track -B <branch> /path/to/dir <remote>/<branch>
+#' ```
+#'
+#' @note this internal function has been modified from the logic in 
+#' [pkgdown::deploy_to_branch()], by Hadley Wickham.
+#'
+#' @keywords internal
+#' @examples
+#' run_ok <- sandpaper:::has_git() &&
+#'   requireNamespace("withr", quietly = TRUE) &&
+#'   rmarkdown::pandoc_available("2.11")
+#'
+#' # Use Worktrees to deploy a lesson -----------------------------------------
+#' # This example is a bit inovlved, but it is effectively what we do inside of
+#' # the `ci_deploy()` function (after setting up the lesson). 
+#' # 
+#' # The setup phase will create a new lesson and a corresponding remote (self
+#' # contained, no GitHub authentication required). 
+#' # 
+#' # The worktrees will be created for both the markdown and HTML outputs on the
+#' # branches "md-outputs" and "gh-pages", respectively. 
+#' # 
+#' # After the worktrees are created, we will build the lesson into the
+#' # worktrees and display the output of `git_status()` for each of the three
+#' # branches: "main", "md-outputs", and "gh-pages"
+#' # 
+#' # During the clean up phase, the output of `git_worktree_setup()` is 
+#' # evaluated
+#' if (run_ok) {
+#' cli::cli_h1("Set up")
+#' cli::cli_h2("Create Lesson")
+#' restore_fixture <- sandpaper:::create_test_lesson()
+#' res <- getOption("sandpaper.test_fixture")
+#' sandpaper:::check_git_user(res)
+#' cli::cli_h2("Create Remote")
+#' rmt <- fs::file_temp(pattern = "REMOTE-")
+#' sandpaper:::setup_local_remote(repo = res, remote = rmt, verbose = FALSE)
+#' cli::cli_h2("Create Worktrees")
+#' db <- sandpaper:::git_worktree_setup(res, fs::path(res, "site", "built"), 
+#'   branch = "md-outputs", remote = "sandpaper-local"
+#' )
+#' ds <- sandpaper:::git_worktree_setup(res, fs::path(res, "site", "docs"), 
+#'   branch = "gh-pages", remote = "sandpaper-local"
+#' )
+#' cli::cli_h1("Build Lesson into worktrees")
+#' build_lesson(res, quiet = TRUE, preview = FALSE)
+#' cli::cli_h2("git status: {gert::git_branch(repo = res)}")
+#' print(gert::git_status(repo = res))
+#' cli::cli_h2('git status: {gert::git_branch(repo = fs::path(res, "site", "built"))}')
+#' print(gert::git_status(repo = fs::path(res, "site", "built")))
+#' cli::cli_h2('git status: {gert::git_branch(repo = fs::path(res, "site", "docs"))}')
+#' print(gert::git_status(repo = fs::path(res, "site", "docs")))
+#' cli::cli_h1("Clean Up")
+#' cli::cli_alert_info("object db is an expression that evaluates to {.code {db}}")
+#' eval(db)
+#' cli::cli_alert_info("object ds is an expression that evaluates to {.code {ds}}")
+#' eval(ds)
+#' sandpaper:::remove_local_remote(repo = res)
+#' sandpaper:::reset_git_user(res)
+#' # remove the test fixture and report
+#' tryCatch(fs::dir_delete(res), error = function() FALSE)
+#' }
+git_worktree_setup <- function (path = ".", dest_dir, branch = "gh-pages", remote = "origin", throwaway = FALSE) {
 
-  # create the branch if it doesn't exist
-  if (no_branch) {
-    old_branch <- gert::git_branch(repo = path)
-    git("checkout", "--orphan", branch)
-    git("rm", "-rf", "--quiet", ".")
-    git("commit", "--allow-empty", "-m", 
-      sprintf("Initializing %s branch", branch)
-    )
-    git("push", remote, paste0("HEAD:", branch))
-    git("checkout", old_branch)
+  if (!has_git() || !requireNamespace("withr", quietly = TRUE)) {
+    stop(cli::format_error("{.fn git_worktree_setup} requires {.pkg git} and {.pkg withr}"), call. = FALSE)
   }
-  # fetch the content of only the branch in question
-  # https://stackoverflow.com/a/62264058/2752888
-  git("remote", "set-branches", remote, branch)
-  git("fetch", remote, branch)
-  github_worktree_add(dest_dir, remote, branch)
+  withr::with_dir(path, {
+    no_branch <- !git_has_remote_branch(remote, branch)
+    # create the branch if it doesn't exist
+    if (no_branch) {
+      old_branch <- gert::git_branch(repo = path)
+      git("checkout", "--orphan", branch)
+      git("rm", "-rf", "--quiet", ".")
+      git("commit", "--allow-empty", "-m", 
+        sprintf("Initializing %s branch", branch)
+      )
+      git("push", remote, paste0("HEAD:", branch))
+      git("checkout", old_branch)
+    }
+    # fetch the content of only the branch in question
+    refspec <- make_refspec(remote, branch)
+    gert::git_fetch(remote = remote, refspec = refspec, repo = path)
+    github_worktree_add(dest_dir, remote, branch, throwaway)
+  })
   # This allows me to evaluate this expression at the top of the calling
   # function.
-  parse(text = paste0("github_worktree_remove('", dest_dir, "')"))
+  parse(text = glue::glue("sandpaper:::github_worktree_remove('{dest_dir}', '{path}')"))
 }
 
 
 # Add a branch to a folder as a worktree
 # originally authored by Hadley Wickham
-github_worktree_add <- function (dir, remote, branch) {
+github_worktree_add <- function (dir, remote, branch, throwaway = FALSE) {
   if (requireNamespace("cli", quietly = TRUE))
     cli::rule("Adding worktree", line = "+")
-  git("worktree", "add", 
-    "--track", "-B", branch, dir, 
-    paste0(remote, "/", branch)
-  )
+  if (throwaway) {
+    the_tree <- c("--detach", dir)
+  } else {
+    the_tree <- c("--track", "-B", branch, dir)
+  }
+  git("worktree", "add", the_tree, paste0(remote, "/", branch))
 }
 
 # Commit on a worktree
@@ -75,7 +171,7 @@ github_worktree_commit <- function (dir, commit_message, remote, branch) {
     # ZNK: Change to gert::git_add(); only commit if we have something to add
     added <- gert::git_add(".", repo = dir)
     if (nrow(added) == 0) {
-      message("nothing to commit!")
+      message(glue::glue("nothing to commit on {branch}!"))
       return(NULL)
     }
     git("commit", "--allow-empty", "-m", commit_message)
@@ -87,11 +183,11 @@ github_worktree_commit <- function (dir, commit_message, remote, branch) {
 
 # Remove a git worktree
 # Modified from pkgdown:::github_worktree_remove by Hadley Wickham
-github_worktree_remove <- function (dir) {
+github_worktree_remove <- function (dir, home = NULL) {
   if (requireNamespace("cli", quietly = TRUE)) 
     cli::rule("Removing worktree", line = "-")
   # ZNK: add --force
-  home <- root_path(dir)
+  if (is.null(home)) home <- root_path(dir)
   if (requireNamespace("withr", quietly = TRUE)) {
     withr::with_dir(home, git("worktree", "remove", "--force", dir))
   }
@@ -118,7 +214,7 @@ message_source <- function(commit_message = "", source_branch = "main", dir = ".
 # - name: "Generate Artifacts"
 #   id: generate-artifacts
 #   run: |
-#     bundle_pr_artifacts(
+#     sandpaper:::ci_bundle_pr_artifacts(
 #       repo         = '${{ github.repository }}',
 #       pr_number    = '${{ github.event.number }}',
 #       path_md      = '${{ env.MD }}',
@@ -127,7 +223,7 @@ message_source <- function(commit_message = "", source_branch = "main", dir = ".
 #       branch       = "md-outputs"
 #     )
 #   shell: Rscript {0}
-bundle_pr_artifacts <- function(repo, pr_number, 
+ci_bundle_pr_artifacts <- function(repo, pr_number, 
   path_md, path_archive, path_pr, 
   branch = "md-outputs") {
   if (!fs::dir_exists(path_archive)) fs::dir_create(path_archive)
@@ -136,41 +232,88 @@ bundle_pr_artifacts <- function(repo, pr_number,
   if (!requireNamespace("withr", quietly = TRUE))
     stop("withr must be installed")
   withr::with_dir(path_md, {
-    git("add", "-A", ".")
+    git("add", "-A", ".", echo_cmd = FALSE, echo = FALSE)
     difflist <- git("diff", "--staged", "--compact-summary",
-      echo = FALSE, echo_cmd = FALSE)
+      echo = FALSE, echo_cmd = FALSE)$stdout
     github_url  <- glue::glue("https://github.com/{repo}/compare/")
-    change_link <- glue::glue("{github_url}{branch}..{branch}-PR-{pr_number}")
-    msg         <- glue::glue(
-      "### Rendered Changes
-
-      :mag: Inspect the changes: {change_link}
-
-      ---
-
-      The following changes were observed in the rendered markdown documents
-
-      ```diff
-      {difflist}
-      ```
-
-      <details>
-      <summary>What does this mean?</summary>
-
-      If you have source files that require output and figures to be generated
-      (e.g. R Markdown), then it is important to make sure the generated 
-      figures and output are reproducible. 
-
-      This output provides a way for you to inspect the output in a 
-      diff-friendly manner so that it's easy to see the changes that occurr due
-      to new software versions or randomisation.
-
-      <details>
-      "
+    reality <- glue::glue("{github_url}{branch}")
+    possibility <- glue::glue("{branch}-PR-{pr_number}")
+    # Comparing commit-ish chunks on GitHub can use either two dot or three dot
+    #
+    # Three dot: compare changes that happened _in that instant_
+    # Two dot: compare the changes between the branches as they exist today.
+    #
+    # https://docs.github.com/en/github/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/about-comparing-branches-in-pull-requests#three-dot-and-two-dot-git-diff-comparisons
+    # 
+    # I was using the two-dot method because that's the only thing I learned, 
+    # but it can be really overwhelming if there are rapid changes. This way,
+    # only the relevant changes are shown (unless there is a conflict).
+    copy_template("pr_diff", path_archive, "diff.md",
+      values = list(
+        reality = reality,
+        possibility = possibility,
+        summary_of_differences = trimws(difflist, which = "right"),
+        update_time = UTC_timestamp(Sys.time()),
+        NULL
+      )
     )
-    writeLines(msg, fs::path(path_archive, "diff.md"))
-    fs::dir_delete(".git")
+    if (fs::is_dir(".git")) fs::dir_delete(".git")
   })
 }
 
+
+# If the git user is not set, we set a temporary one, note that this is paired
+# with reset_git_user()
+check_git_user <- function(path) {
+  if (!gert::user_is_configured(path)) {
+    gert::git_config_set("user.name", "carpenter", repo = path)
+    gert::git_config_set("user.email", "team@carpentries.org", repo = path)
+  }
+}
+
+# It's clear that we cannot rely on folks having the correct libgit2 version,
+# so the way we enforce the main branch is to do it after we make the initial
+# commit like so:
+#
+#  1. create a new branch called "main"
+#  2. change "master" to "main" in .git/HEAD (txt file)
+#  3. delete "master" branch 
+# 
+# If the user HAS set a default branch, we will use that one.
+enforce_main_branch <- function(path) {
+  current <- gert::git_branch(path)
+  default <- get_default_branch()
+  if (current != "master") {
+    # the user set up their init.defaultBranch correctly
+    return(path)
+  }
+  # Create and move to main branch
+  gert::git_branch_create(default, repo = path)
+  # modify .git/HEAD file
+  HFILE <- file.path(path, ".git", "HEAD")
+  HEAD <- readLines(HFILE, encoding = "UTF-8")
+  writeLines(sub("master", default, HEAD), HFILE)
+  # remove master
+  gert::git_branch_delete("master", repo = path)
+}
+
+get_default_branch <- function() {
+  cfg <- gert::git_config_global()
+  default <- cfg$value[cfg$name == "init.defaultbranch"]
+  invalid <- length(default) == 0 || default == "master"
+  if (invalid) "main" else default
+}
+
+# This checks if we have set a temporary git user and then unsets it. It will 
+# supriously unset a user if they happened to have 
+# "carpenter <team@carpentries.org>" as their email.
+reset_git_user <- function(path) {
+  cfg <- gert::git_config(path)
+  it_me <- cfg$value[cfg$name == "user.name"] == "carpenter" &&
+    cfg$value[cfg$name == "user.email"] == "team@carpentries.org"
+  if (gert::user_is_configured(path) && it_me) {
+    gert::git_config_set("user.name", NULL, repo = path)
+    gert::git_config_set("user.email", NULL, repo = path)
+  }
+}